backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.25
google-api-python-client
google-genai